entrypoint: installing devel libraries is now faster when using supported dtprojects

+++ assets/entrypoint.sh
@@ -282,17 +282,26 @@ configure_libraries() {
                     if [ -f "${LIBRARY_SETUP_PY}" ] || [ -f "${LIBRARY_PYPROJECT_TOML}" ]; then
                         debug "  > Found library in ${lib}"
                         # ---
-                        # copy the library to a temporary directory
-                        LIB_NAME=$(basename ${lib})
-                        TMP_LIB_DIR=$(mktemp -d)
-                        cp -R "${lib}" "${TMP_LIB_DIR}"
-                        # poetry expects a full repository while libraries are submodules with a fake .git file
-                        export POETRY_DYNAMIC_VERSIONING_BYPASS="0.0.0"
-                        export POETRY_DYNAMIC_VERSIONING_COMMANDS=""
-                        # install the library
-                        python3 -m pip install --no-dependencies --editable "${TMP_LIB_DIR}/${LIB_NAME}" > /dev/null
+                        if [[ ( -d "${lib}/dtproject" || -f "${lib}/.dtproject" ) && -d "${lib}/src" ]]; then
+                            # we recognize the library as a dtproject, use the PYTHONPATH trick
+                            info "  > Adding ${lib}/src to PYTHONPATH"
+                            export PYTHONPATH="${lib}/src:${PYTHONPATH}"
+                            info "  < Loaded library: ${LIB_NAME} (exposed to PYTHONPATH)"
+                        else
+                            # we don't recognize the library, install it using pip instead
+                            info "  > Installing ${lib} using pip"
+                            # copy the library to a temporary directory
+                            LIB_NAME=$(basename ${lib})
+                            TMP_LIB_DIR=$(mktemp -d)
+                            cp -R "${lib}" "${TMP_LIB_DIR}"
+                            # poetry expects a full repository while libraries are submodules with a fake .git file
+                            export POETRY_DYNAMIC_VERSIONING_BYPASS="0.0.0"
+                            export POETRY_DYNAMIC_VERSIONING_COMMANDS=""
+                            # install the library
+                            python3 -m pip install --no-dependencies --editable "${TMP_LIB_DIR}/${LIB_NAME}" > /dev/null
+                            info "  < Loaded library: ${LIB_NAME}\t(installed from temporary directory: ${TMP_LIB_DIR})"
+                        fi
                         # ---
-                        info "  < Loaded library: ${LIB_NAME}\t(from temporary directory: ${TMP_LIB_DIR})"
                     fi
                 done
                 debug " > Analyzed ${PROJECT_LIBRARIES_DIR}/"
